add ClientHandlerUtils for helper functions
